Part 4.21 done

+++ part4/bloglist/controllers/blogs.js
@@ -47,8 +47,20 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+    const userId = await User.findById(decodedToken.id)
+
+    if (blog.user.toString() === userId._id.toString()) {
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+    } else {
+        response.status(401).json({ error: 'invalid user' })
+    }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
